Make information paramater 32 bit

+++ cpp/base/primary_device.h
@@ -132,7 +132,7 @@ protected:
     void SetFilemark();
     void SetEom(ascq);
     void SetIli();
-    void SetInformation(int64_t);
+    void SetInformation(int32_t);
 
     void StatusPhase() const;
     void DataInPhase(int) const;
